+++ src/input/mod.rs
@@ -1,3 +1,7 @@
+pub mod event;
+
+use std::fmt::{self, Display};
+
 use crate::model::Model;
 
 pub enum Message {
@@ -25,16 +29,16 @@ pub fn update(model: &mut Model, msg: Message) -> Option<Message> {
     None
 }
 
-// impl Display for Message {
-//     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-//         let str = match self {
-//             Message::RowUp => "Move One Row Up",
-//             Message::RowDown => "Move One Row Down",
-//             Message::Reset => "Go To Top Row",
-//             Message::BottomRow => "Go To Bottom Row",
-//             Message::Nothing => "",
-//             Message::Quit => "Quit",
-//         };
-//         write!(f, "{}", str)
-//     }
-// }
+impl Display for Message {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let str = match self {
+            Message::RowUp => "Move One Row Up",
+            Message::RowDown => "Move One Row Down",
+            Message::Reset => "Go To Top Row",
+            Message::BottomRow => "Go To Bottom Row",
+            Message::Nothing => "",
+            Message::Quit => "Quit",
+        };
+        write!(f, "{}", str)
+    }
+}
