+++ src/utils/cli.rs
@@ -3,11 +3,13 @@ use clap::Parser;
 #[derive(Parser, Default, Debug)]
 #[clap(author = "Ddraigan", version, about = "A side by side git diff view")]
 pub struct Arguments {
-    #[clap(short = 'C', long, action)]
+    #[clap(short = 'C', long)]
     /// Giving a full path allows diff-tool to diff outside of the git repo
     change_dir: bool,
     /// File to diff with
     path: String,
+    #[clap(short, long, default_value_t = 250)]
+    tick_rate: u64,
 }
 
 impl Arguments {
@@ -17,5 +19,9 @@ impl Arguments {
 
     pub fn change_dir(&self) -> bool {
         self.change_dir
+    }
+
+    pub fn tick_rate(&self) -> u64 {
+        self.tick_rate
     }
 }
